# openapi.yaml
...
openapi: 3.0.1
info:
  title: Dataspace Connector
  description: IDS Connector reference implementation
  contact:
    name: Dataspace Connector
    url: https://github.com/International-Data-Spaces-Association/DataspaceConnector
    email: support@sovity.de
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
  version: 8.0.0
servers:
  - url: https://localhost:8080
    description: Generated server url
security:
  - basicAuth: []
...
paths:
...
  /api/examples/validation:
    post:
      tags:
        - _Utils
      summary: Get the policy pattern represented by a given JSON string.
      operationId: getPolicyPattern
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: The JSON string representing a policy.
        required: true
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
  /api/examples/policy:
    post:
      tags:
        - _Utils
      summary: Get an example policy for a given policy pattern.
      operationId: getExampleUsagePolicy
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/PatternDesc'
                - $ref: '#/components/schemas/ConnectorRestrictionDesc'
                - $ref: '#/components/schemas/DeletionDesc'
                - $ref: '#/components/schemas/DurationDesc'
                - $ref: '#/components/schemas/IntervalDesc'
                - $ref: '#/components/schemas/LoggingDesc'
                - $ref: '#/components/schemas/NotificationDesc'
                - $ref: '#/components/schemas/PermissionDesc'
                - $ref: '#/components/schemas/ProhibitionDesc'
                - $ref: '#/components/schemas/SecurityRestrictionDesc'
                - $ref: '#/components/schemas/UsageNumberDesc'
        required: true
      responses:
        "200":
          description: Ok
          content:
            application/ld+json:
              schema:
                type: object
        "400":
          description: Bad request
          content:
            application/ld+json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/ld+json:
              schema:
                type: object
...
components:
  schemas:
...
    ConnectorRestrictionDesc:
      type: object
      example:
        type: CONNECTOR_RESTRICTED_USAGE
        url: https://localhost:8080
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
        - type: object
          properties:
            url:
              type: string
    DeletionDesc:
      type: object
      example:
        type: USAGE_UNTIL_DELETION
        start: 2020-07-11T00:00:00Z
        end: 2020-07-11T00:00:00Z
        date: 2020-07-11T00:00:00Z
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
        - type: object
          properties:
            start:
              type: string
            end:
              type: string
            date:
              type: string
    DurationDesc:
      type: object
      example:
        type: DURATION_USAGE
        duration: PT1M30.5S
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
        - type: object
          properties:
            duration:
              type: string
    IntervalDesc:
      type: object
      example:
        type: USAGE_DURING_INTERVAL
        start: 2020-07-11T00:00:00Z
        end: 2020-07-11T00:00:00Z
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
        - type: object
          properties:
            start:
              type: string
            end:
              type: string
    LoggingDesc:
      type: object
      example:
        type: USAGE_LOGGING
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
    NotificationDesc:
      type: object
      example:
        type: USAGE_NOTIFICATION
        url: https://localhost:8080/api/ids/data
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
        - type: object
          properties:
            url:
              type: string
    PatternDesc:
      required:
        - type
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
      discriminator:
        propertyName: type
    PermissionDesc:
      type: object
      example:
        type: PROVIDE_ACCESS
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
    ProhibitionDesc:
      type: object
      example:
        type: PROHIBIT_ACCESS
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
    SecurityRestrictionDesc:
      type: object
      description: "Possible profiles are: idsc:BASE_SECURITY_PROFILE, idsc:TRUST_SECURITY_PROFILE\
        \ and idsc:TRUST_PLUS_SECURITY_PROFILE"
      example:
        type: SECURITY_PROFILE_RESTRICTED_USAGE
        profile: BASE_SECURITY_PROFILE
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
        - type: object
          properties:
            profile:
              type: string
    UsageNumberDesc:
      type: object
      example:
        type: N_TIMES_USAGE
        value: "5"
      allOf:
        - $ref: '#/components/schemas/PatternDesc'
        - type: object
          properties:
            value:
              type: string
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic